+++ test/util.jl
@@ -9,7 +9,7 @@ spectrum_func2(D=10) = spectrum(ϵ->f(D, ϵ), lb=0, ub=D) # half-line spectrum
 
 # ------------------------------------------------------- Grassmann orderings
 const imag_orderings   = [A1Ā1B1B̄1(), A1B1B̄1Ā1()]
-const real_orderings   = [A1Ā1B1B̄1a1ā1b1b̄1(), A1Ā1a1ā1B1B̄1b1b̄1(),
+const real_orderings   = [A1Ā1B1B̄1a1ā1b1b̄1(), A1Ā1a1ā1B1B̄1b1b̄1(), a1ā1A1Ā1b1b̄1B1B̄1a2ā2A2Ā2b2b̄2B2B̄2(),
 						  A1B1ā1b̄1Ā1B̄1a1b1(),
 						  A2B2B̄2Ā2A1B1B̄1Ā1a1b1b̄1ā1a2b2b̄2ā2(),
 						  A2Ā2B2B̄2A1Ā1B1B̄1a1ā1b1b̄1a2ā2b2b̄2()]